+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 0.7.1-xx-evcommon-dev
+
+- All iOS operations are now using background processing (reported to resolve intermittent problems with cordova-ios@4.0.1)
+
 ## 0.7.1-evcommon-dev
 
 - Rename Windows C++ Database close function to closedb to resolve conflict for Windows Store certification